refactor(ui-document): extract completion helpers to completion_support module
