+++ crates/dbflux_ui_document/src/lib.rs
@@ -2,6 +2,7 @@
 #![allow(unused_imports)]
 
 mod add_member_modal;
+pub(crate) mod completion_support;
 mod audit;
 pub mod chart;
 pub mod chart_document;
